backend/requirements.txt
# SkinCancerDetector - Python Dependencies
# Install: pip install -r requirements.txt

# Core ML
# NOTE:
# - The API can run in DEMO mode without TensorFlow (see models/model_architecture.py).
# - TensorFlow 2.15.0 does not support Python 3.12. We guard it with an env marker to
#   avoid hard install failures on modern Python.
tensorflow==2.15.0; python_version < "3.12"
keras==2.15.0; python_version < "3.12"

# Data Processing  
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.3.0

# Image Processing
Pillow==10.1.0
# 4.8.1 is not a valid PyPI release tag; use the full build version.
opencv-python==4.8.1.78

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Optional (for visualization)
matplotlib==3.8.0
seaborn==0.13.0
